Export LLMSession and SyncSession to make it easier to add support for additional models
Hi @slundberg and all, first off great work on this project, I'm very excited to see how it develops.

As per #58 it would be very useful to be able to extend guidance to support additional LLM models and make use of all of the features.

I understand this project is quite new and you probably want to avoid the cost of maintaining N implementations. If it's an acceptable solution, can we export `LLMSession` and `SyncSession` so external projects can add support on their own?