Grammar Builder for JSON Schema (#619)
A very basic `guidance` grammar builder for JSON schema. Handles the basic types, although the set of valid strings is more restricted than JSON (because making new things acceptable is preferable to rejecting things previously allowed).

There is no support for cross references; that will have to wait until a future PR. There is also no support for constraints. I do not expect constraints such as `multipleOf` or `exclusiveMinimum` to be supportable. It is possible that some of the constraints such as `minItems`, `maxLength` and `pattern` may be supportable.