Separate out server tests (#688)
The tests in `test_server.py` seem to be a touch temperamental, often failing when the forked server process fails to start in time. In an effort to reduce the re-running of tests, add a `server` mark for these tests, and run in a separate runner job. Use a new workflow file so that everything doesn't get killed by the first failure.