[Test] Enable Azure AI notebook (#721)
Start running the Azure AI notebook in tests. This is causing a rethink of some of the testing infrastructure, to add a new `needs_credentials` tag for selecting tests which require API keys and the like.

A fixture for rate limiting is also required, since otherwise we overwhelm our endpoints.