replace catches of the ImportError exception with the more specific ModuleNotFoundError exception (#733)
This should partly address
https://github.com/guidance-ai/guidance/issues/690 where it appears the
import of llama-cpp is generating an ImportError exception that is not
due to the module not being installed. This more specific exception will
only catch the specific case of the Module not being installed and will
leave other exceptions to propagate to the user with more descriptive
errors.