fix for issue #660 where the call to llama_batch_free is failing during process shutdown (#734)
llama_batch_free is failing during python shutdown when the interpreter
can set objects to None non-deterministically. I believe Guidance is
properly checking for None in the objects that can be accessed from
Guidance, but llama-cpp-python is not, and the native library object in
llama-cpp-python is being set to None before it can be called. Since
we're only dealing with memory deallocation here, and since the process
is being killed anyway, we can solve this on the Guidance side by not
calling llama_batch_free during process termination.