[Test] Fix for JSON tests (#797)
We have been getting intermittent failures in the JSON tests. These are occurring in the 'number' and 'integer' tests, and are almost certainly due to the behaviour of the `model.Mock` class.

Once its supplied string has been exhausted, `Mock` generates random characters (or to be more precise, produces randomised logits for the valid tokens). This means that every now and then, a digit can be produced, which will pass the grammar check for a number, but then fail our round-trip check. This is not a problem for any of our other tests, since in those cases, the final character is forced (e.g. `"` to close a string or `]` to end an array), and there is no possible valid character afterwards.

The fix is to append a character to the string supplied to `Mock` which _never_ appears in the one we supply from the test. This acts as a universal 'stop code.'