[Build] Separate notebooks testing (#825)
The tests of our notebooks (currently running in the CI build) appear to be particularly unreliable, and also very slow. This is likely not unrelated to the large number of model calls which they make. For now, split them into a CI build of their own, so failures in the 'real' CI build are more obvious.