[Test] Environment variable for selected model (#883)
Make debugging tests with a particular LLM easier by adding a `GUIDANCE_SELECTED_MODEL` environment variable.

We have the `--selected_model` command line argument for `pytest`, which loads a designated LLM in a fixture (if not set, currently `gpt2cpu` is the default). However, when debugging tests within VS Code, setting that command line argument is non-trivial, and our workaround was to patch the `selected_model_name()` fixture (which reads the command line argument). This change makes the default value for the command line argument be the `GUIDANCE_SELECTED_MODEL` environment variable, and if _that_ is not set, falls back to `gpt2cpu`.