[Bug] Passing through kwargs to `TransformersTokenizer` (#922)
This is a fix for #909 . It ensures that any argument common to both `AutoModelForCausalLM.from_pretrained()` and `AutoTokenizer.from_pretrained()` will be passed from the former to the latter when a `TransformersModel` is created. The original issue was only about `trust_remote_code` but there are a few other arguments related to accessing Hugging Face which probably ought to be forwarded as well.

If there are LLM specific arguments which need to be shared, then users should instantiate the tokeniser separately, and pass it in.