[Feature] Add extra support to JSON string schema (#927)
Looking to add support for `pattern`, `minLength` and `maxLength` to strings in JSON schema. This should address half of #925

This has lead to the creation of a few new library functions to help out, and consolidation into a new `_sequences.py` library file. It is also hooked into `_regex.py`.