[Build] Fix CI and Notebook builds (#940)
Get the CI and Notebook builds working again. The main fix involves the authentication to the AzureAI endpoints; we now have an Entra ID to use, rather than an `api_key`. The instructions are [on the Microsoft site](https://learn.microsoft.com/en-us/azure/developer/github/connect-from-azure).

Also update the chat templates again, for Phi3 and Mistral. The fix for Mistral is not ideal. It appears that it now 'supports' a `system` message by inserting it into the first `user` message. I don't think our approach can really support this.