[Build] Fix CI build (Llama2 Chat Templates) (#948)
A couple of issues in the CI build were not caught because the PR was from a fork. There are problems with the Llama2 chat templates, which are addressed here. There are some fixes to the templates themselves, but the handling of the `system` role is so different from how the Guidance context managers work that the test is XFAILed.