[Bug] Fix regex for urls in image function (#928)
The regex for matching urls in the `image` function starts with a `$` character ([ref](https://github.com/guidance-ai/guidance/blob/ba86754265f2ca1c772c6df1d9713df42c7d704a/guidance/library/_image.py#L13)), i.e. the "match end of string"-metacharacter, which seems to be wrong. I guess the intention was to match against the start of the string, which could be achieved with a `^` character, but since `re.match` matches against the start of a string already, this can also be omitted in this specific case.

This led to the situation that urls are generally not recognized and can't be used with the `image` function. This PR attempts to fix this issue by modifying the aforementioned regex and additionally providing tests for the `_image` module to prevent such problems in the future.