[Feature] Llama3 Support (#954)
Ensure that Llama3 works with Guidance. Since this is a 'protected' model, the testing has to be run in the CI build, so we have access to the `HF_TOKEN`. Due to limitations on the build machines, the testing is minimal, unfortunately. There are also a few quality-of-life improvements to the test setup.