Remove `Placeholder` and `replace_grammar_node` in favor of pointer-like container object (#1007)
Essentially reopening #638 with some major simplifications allowed by
the rust re-write (the serialized grammar now directly supports
"references").

Reopening because:
#995 still takes far too long to process the large JSON schema.
Line-profiling revealed that ~80% of the time spent constructing the
`GrammarFunction` is due to `replace_grammar_node`. In particular, this
function has to traverse the entire grammar tree, and we potentially
have to call it many times if there is a lot of mutual recursion.

Simply adding a node type that acts as a container (which we can fill
after we decide what its contents should be) side-steps this problem
entirely.