[Feature] Enable user-passed `separators` keyword to `guidance.json` (#1044)
Main changes:
1. Refactor json generation code to live in a class with
guidance-decorated methods (allows separators to be set at the instance
level rather than having to pass them to every function call; note that
`definitions` are now also stored on the instance)
2. Add `separators` kwarg that controls the strings we will use for item
and key separators (`", "` and `": "` by default). This mirrors the
`json.dumps` API.
3. Add `whitespace_flexible` kwarg to enable old default behavior of
"letting the LLM decide" on whitespace. This is no longer the default
behavior, as it is substantially less efficient from a
guidance-acceleration standpoint.
4. Deprecate `compact` kwarg in favor of `separators = (",", ":")`
5. Proper handling of `const` and `enum` types to ensure correct lexical
boundaries (these were previously incorrectly handled from the point of
view of whitespace flexibility).
6. Change existing tests to reflect the new default behavior
7. Add explicit whitespace tests for cases where `separators` or
`whitespace_flexible` are passed