Model/State/Client Architecture + OpenAI Audio Outputs (#1136)
This PR adds support for `gen_audio` (potentially not the final public
API) for OpenAI audio-enabled models.

Note that this involves a pretty significant refactor of the initial
implementation of "State" objects introduced a PR or two ago and
centralizes handling of different input/output formats in `Client`
classes. They may need to be renamed now that they do more heavy
lifting. But in all, I think this is far more flexible and easy to
extend.