[Hotfix] Only infer client_cls from model when it's a str (#1165)
Currently, `main` does not support passing `AutoTokenizer`/`AutoModel`
to the `Transformers` class (e.g. see
https://github.com/guidance-ai/guidance/issues/1144).

This is because the `Client` class expected `model` to be passed as a
string for the sake of inferring some capabilities of the model. Now
there is an explicit `isinstance` check, as well as some additional type
annotations to help prevent devs (me) from having the wrong mental model
about types here.

For future: we should expand the capabilities check to be able to
introspect on `AutoModel` classes / config files as well as looking at
the raw string. Not in the scope of this PR though.