[FEAT] Azure AI Update (#1184)
A basic update to support for Azure AI models. Credit to @hudson-ai for doing the majority of the work creating the new `Interpreter` classes.

There are complications arising from the fact that Azure AI has two separate APIs, dependent on whether one is working with an OpenAI model, or with one of the other available models. Furthermore, the two sides work slightly differently.

On the OpenAI side, the endpoint (e.g. `https://aoai-guidance.openai.azure.com/`), deployment name (e.g.
`guidance-gpt-4o-mini`) and model name (e.g. `gpt-4o-mini`) are distinct. The start of the endpoint is the name of the "Azure OpenAI" resource in the Azure portal itself (as distinct from the Azure AI portal).

On the Other Models side, the deployment name gets put into the endpoint (e.g. `https://guidance-phi4.swedencentral.models.ai.azure.com`) while the model name remains separate (e.g. `Phi-4`). The name of the project within the Azure AI Foundry, or the name of the Azure AI Foundry itself does not appear.

---------

Co-authored-by: Hudson Cooper <hncooper96@gmail.com>