[Refactor] Use `llguidance`'s native wrapper for `transformers.PretrainedTokenizerFast` where possible (#1214)
1. Refactors all tokenizers to wrap `llguidance.LLTokenizer` rather than
constructing an `LLTokenizer` on the fly when we run the parser
2. Doing so allows each `Tokenizer` class to have control over how it
constructs said `LLTokenizer`.
- `TransformersTokenizer` is able to call `llguidance.hf.from_tokenizer`
directly if we have a `transformers.PretrainedTokenizerFast`, bypassing
the slow and error-prone preprocessing step of inferring `byte_tokens`
- `TransformersTokenizer` falls back to the old approach for non-fast
tokenizers
- `LlamaCppTokenizer`, `MockTokenizer`, and `ByteTokenizer` all use the
old approach (manually specifying a list of `byte_tokens`). They also
implement their own `encode` methods, which are notably passed to the
`LLTokenizer` that they wrap

Note that we may get a similar `llguidance.llamacpp.from_tokenizer` for
a "native" cpp tokenizer at some point :)

P.S. this refactor may indicate that we actually want to get rid of the
`Tokenizer` class all-together, simply using `LLTokenizer` in the
`Engine`s. That felt a bit extreme at the moment though, so I decided to
leave it as-is.