[Build] Work around sentencepiece (#1223)
It appears that the current release of `sentencepiece` which is available on PyPI does not support Python 3.13. So for this specific version, use a community provided build for 3.13. See issue #1222 for further details.

This fix will break if `sentencepiece` still hasn't been updated by the time we try moving to Python 3.14 (and if this is not straightened out when we move to Python 4.13, a major version check will need to be added). This is intentional; a fix this ugly should be somewhat fragile, so that we don't forget it's there.