Add probabilities back in for ff_tokens (#1261)
Computes the probabilities of fast-forwarded tokens (in addition to
generated tokens, for which we already had probabilities).

To implement this, I implemented a "logit cache" for the
`TransformersEngine` and `LlamacppEngine` classes, which maintains a
cached version of all `(seq_len, vocab_size)` logits that have been
computed thus far. Each call into the underlying model object will
(possibly) truncate and then append to the cache (i.e. it has the same
semantics as the KV cache, but its size should be pretty negligable in
comparison).

There is a slight downside -- we need to call the model one extra time
at the end of generation IF the generation ended with fast-forwarded
tokens. We need to do this in order to get their probabilities. Another
note here, but I'm currently caching the logits as numpy arrays. In
principle, it might be nice to keep torch tensors on their original
device... All of the round trips might be kind of expensive. That being
said, this issue was already present in the previous implementation.

It would be great to follow this PR up with something that gets the
top-k tokens for ff tokens too.