fix bug that was skewing token probabilities when masking was applied (#1266)
Previously, the logits were skewed substantially when masking was
applied since the values inside the mask variable were either 0 or 200
depending on whether they were allowed or not allowed. The value of 200
was multiplied by the logits, which skewed them much higher than
intended, leading to the top token most likely always being selected
instead of probabilistically selecting them. A further fix to
temperature is included. Previously temperatures of 0.0 were treated the
same as temperature = 1.0 since softmax was used on the original logits
instead of deterministically selecting the highest logit token.