[Hotfix] Fix performance regression introduced in #1261 (#1276)
#1261 added a logit cache, which was being concatenated to in a loop.

This PR never caches any logits except for the final one, but if
`include_all_uncached_tokens` is `True`, then `get_logits` will return
an array of shape `(len(tokens)-num_cached, vocab_size)`. Otherwise, it
will return an array of shape `(1, vocab_size)`

This "should" include all tokens being fast-forwarded (unless the
fast-forwarded tokens are somehow already in the kv, e.g. if a notebook
cell is re-run). If any are missing, I just call the probabilities `nan`
here, which could in principle cause some artifacts in the widget.

Note that an alternative implementation would have been to keep the full
logit cache, pre-allocating the buffer in advance with shape `(n_ctx,
vocab_size)`, but this feels really unnecessary. Happy to revisit after
we get this merged and the big performance issues are gone.

@nopdive would you mind re-running the repro? Note that you shouldn't
see a speedup for the prompt part (as this PR does not concern
pre-filling), but you should see a speedup for the rest of the
completion. Side note, but we can get rid of the (relatively small, mind
you) prompt slowdown by simply deciding to omit probabilities from
"input" tokens. Lmk if you want that, and I can open a follow-up PR.

Edit:
Intended to fix #1275