[Bug] Fix AssertionError with Gemma2 (#1277)
Gemma2 has a non-croppable cache, so the cache is reset every time we do
a forward pass that doesn't simply append tokens. In this case,
`num_cached` wasn't being set to 0 like `past_length`, causing a later
assertion to fail.

Note that this fix does not change any behaviors, as `num_cached` isn't
actually used for anything after this line (except for the assertion).