[BUG] Terminal rule indirection (#1329)
When a rule would be allowed to be terminal if not for its attributes
(i.e. its RHS consists only of concatenations and alternations of
terminals, but it has an attribute like `stop` or `capture`), add an
extra *terminal* rule at serialization time that sits between the
original attributed rule and its RHS.

This ensures that the addition of attributes doesn't artificially
prevent us from grouping items into a single terminal/lexeme. Maximising
performance means reducing the number of individual lexemes - indeed,
llguidance imposes a limit on the total number of lexemes for this very
reason.

Should fix #1320

---------

Co-authored-by: Richard Edgar (Microsoft) <riedgar@microsoft.com>