[Bug] Normalised capture names not preserved (#1334)
We had a bug in the processing of capture names. If a capture name is also a valid Lark rule name, then it can be left in place. Otherwise (e.g. if it mixes upper and lower case), we have to normalise it, and use the `capture="OriginalName"` feature of our Lark grammar. This wasn't being done.

Fixes #1172

---------

Co-authored-by: Hudson Cooper <hncooper96@gmail.com>