[Draft] Add `chat_completion()` to the `Engine` (#1345)
Add two experimental APIs to the `Engine` class, to provide chat completion abilities similar to the OpenAI API. The goal is to allow users to constrain generations from a local LLM using a Lark grammar, without using the DSL we have built in Python. The added methods are `chat_completion()` and `chat_completion_streaming()`.

When using this, users will have to manage the conversation state themselves.