[Build] Fixes (#1382)
The Docs build was running out of disk space, so move to `Large_Linux` nodes (not sure what caused this, but it's not a high priority to run down).

Also fix a small issue in the Notebooks build. Since there are now two `pytest` invocations, we need to make sure that the second doesn't overwrite coverage results from the first.