[Bug] Fix chat_completion (#1387)
Turns out that there's an argument `add_generation_prompt` in the chat templates (for Hugging Face at least), and it's really wiser to use this than trying to do things manually.

Caught this while messing around with other grammars - Phi4 was reluctant to generate regular text (instead of a tool call), apparently because an `<|endoftext|>` token was present in the prompt (because `add_generation_prompt` had defaulted to false).

Also fixes `ruff` error on a `test_notebooks_azure.py` because that's not worth a separate PR.